requirements.txt
torch>=1.3.1
pillow>=5.4.1
numpy>=1.16.2
torchvision>=0.4.2
PyYAML>=5.1
tensorboardX>=1.7
opencv-python
tensorboard
cython
imageio>=2.5.0
scikit-image>=0.15.0
chainercv>=0.12.0
git+https://github.com/lucasb-eyer/pydensecrf.git
git+https://github.com/cocodataset/cocoapi.git#subdirectory=PythonAPI